refactor: add dropdown menu component to storybook
