+++ src/components/community/article/ArticleSection/index.tsx
@@ -1,7 +1,7 @@
 import DOMPurify from "isomorphic-dompurify";
 import CommonInfo from "@components/community/CommonInfo";
 import { ArticleResponse } from "@lib/API/petBookAPI/types/articleRequest";
-import DropdownMenu from "@components/common/DropdownMenu";
+import DropdownMenu from "@/stories/common/DropdownMenu";
 import { BookmarkBlankIcon } from "@components/common/icon/Bookmark";
 import useModal from "@lib/hooks/common/useModal";
 import { articleRequest } from "@lib/API/petBookAPI";
@@ -17,6 +17,7 @@ const dummyImage =
 
 const ArticleSection = ({ data }: { data: ArticleResponse | undefined }) => {
   const { userData } = useUserInfo();
+  console.log(userData);
   const { openModal, closeModal } = useModal();
   if (data === undefined) {
     return <ArticleSectionBox />;
@@ -58,7 +59,7 @@ const ArticleSection = ({ data }: { data: ArticleResponse | undefined }) => {
     <ArticleSectionBox>
       <div className="ArticleSection_Top_Row">
         <h2>{title}</h2>
-        {user.id === userData?.id && <DropdownMenu menuList={menuList} />}
+        {user.id == userData?.id && <DropdownMenu menuList={menuList} />}
       </div>
       <CommonInfo
         avatar={dummyImage}
