+++ src/stories/common/DropdownMenu/index.tsx
@@ -1,7 +1,7 @@
-import React, { MouseEventHandler, useState } from "react";
-import styled from "styled-components";
-import ThreeDotsVerticalIcon from "./icon/ThreeDotsVertical";
-import OnClickOutside from "./OnClickOutside";
+import React, { type MouseEventHandler, useState } from "react";
+import ThreeDotsVerticalIcon from "../../assets/ThreeDotsVertical";
+import OnClickOutside from "../OnClickOutside";
+import { MenuBox, MenuButton, MenuList } from "./DropdownMenu.styled";
 
 interface Props {
   menuList: {
@@ -33,37 +33,5 @@ const DropdownMenu = ({ menuList }: Props) => {
     </OnClickOutside>
   );
 };
-
-const MenuBox = styled.div`
-  position: relative;
-  width: 24px;
-  height: 24px;
-`;
-
-const MenuButton = styled.button`
-  width: 100%;
-  height: 100%;
-  color: #92928e;
-`;
-
-const MenuList = styled.div`
-  position: absolute;
-  top: 24px;
-  right: 0;
-  width: 100px;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 16px;
-  padding: 16px 0 16px 20px;
-  background-color: white;
-  border: 1px solid var(--bg_white_01);
-  box-shadow: 0px 8px 12px rgba(50, 26, 9, 0.1);
-  border-radius: 12px;
-  button {
-    width: 100%;
-    text-align: left;
-  }
-`;
 
 export default DropdownMenu;
